bindgen: fix build
The bindgen command was causing a panic without this change. We should
not be coupling bindgen to op-e2e. There is no reason to trigger the
init function inside of op-e2e when running bindgen. This commit
dedupes the `FindMonoRepoRoot` function by putting it into a common
location where it can be consumed. Not bindgen no longer depends on
op-e2e. The op-e2e package is for tests, not defining library functions
that are used elsewhere in the repo.
