+++ indexer/config/devnet.go
@@ -3,10 +3,11 @@ package config
 import (
 	"encoding/json"
 	"errors"
-	"fmt"
 	"io/fs"
 	"os"
 	"path/filepath"
+
+	op_service "github.com/ethereum-optimism/optimism/op-service"
 )
 
 var DevnetPresetId = 901
@@ -17,7 +18,7 @@ func DevnetPreset() (*Preset, error) {
 		return nil, err
 	}
 
-	root, err := findMonorepoRoot(cwd)
+	root, err := op_service.FindMonorepoRoot(cwd)
 	if err != nil {
 		return nil, err
 	}
@@ -41,26 +42,4 @@ func DevnetPreset() (*Preset, error) {
 		Name:        "Local Devnet",
 		ChainConfig: ChainConfig{Preset: DevnetPresetId, L1Contracts: l1Contracts},
 	}, nil
-}
-
-// findMonorepoRoot will recursively search upwards for a go.mod file.
-// This depends on the structure of the monorepo having a go.mod file at the root.
-func findMonorepoRoot(startDir string) (string, error) {
-	dir, err := filepath.Abs(startDir)
-	if err != nil {
-		return "", err
-	}
-	for {
-		modulePath := filepath.Join(dir, "go.mod")
-		if _, err := os.Stat(modulePath); err == nil {
-			return dir, nil
-		}
-		parentDir := filepath.Dir(dir)
-		// Check if we reached the filesystem root
-		if parentDir == dir {
-			break
-		}
-		dir = parentDir
-	}
-	return "", fmt.Errorf("monorepo root not found")
 }
